Move core.py, utilities.py to lib/
