+++ tests/__init__.py
@@ -1,4 +1,6 @@
 import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
 ASSETS_PATH = os.path.join(os.path.dirname(__file__), 'assets')
 REPOS_PATH = os.path.join(ASSETS_PATH, 'repos')
